first run at vim.pack
