+++ nvim/.config/nvim/after/ftplugin/csv.lua
@@ -1,0 +1,1 @@
+require("csvview").enable()
