+++ nvim/.config/nvim/after/ftplugin/tsv.lua
@@ -1,0 +1,1 @@
+require("csvview").enable()
