+++ nvim/.config/nvim/init.lua
@@ -1,12 +1,4 @@
 vim.loader.enable()
 require("config.options")
----@diagnostic disable-next-line: different-requires
-require("config.lazy")
-
-vim.api.nvim_create_autocmd("User", {
-  pattern = "VeryLazy",
-  callback = function()
-    require("config.autocmds")
-    require("config.keymaps")
-  end,
-})
+require("config.autocmds")
+require("config.keymaps")
