--- nvim/.config/nvim/lua/plugins/lsp.lua
@@ -1,242 +1,0 @@
-local icons = require("config.icons")
-
--- diagnostics
-vim.diagnostic.config({
-  severity_sort = true,
-  signs = {
-    text = {
-      [vim.diagnostic.severity.ERROR] = icons.diagnostics.Error,
-      [vim.diagnostic.severity.WARN] = icons.diagnostics.Warn,
-      [vim.diagnostic.severity.HINT] = icons.diagnostics.Hint,
-      [vim.diagnostic.severity.INFO] = icons.diagnostics.Info,
-    },
-  },
-  virtual_text = {
-    source = "if_many",
-    prefix = "",
-    severity = {
-      min = vim.diagnostic.severity.ERROR,
-    },
-  },
-  underline = {
-    severity = {
-      min = vim.diagnostic.severity.HINT,
-    },
-  },
-  jump = {
-    -- use virtual lines for jumps
-    -- help diagnostic-on-jump-example
-    on_jump = function(diagnostic, bufnr)
-      if not diagnostic then
-        return
-      end
-      vim.diagnostic.show(
-        diagnostic.namespace,
-        bufnr,
-        { diagnostic },
-        { virtual_lines = { current_line = true }, virtual_text = false }
-      )
-    end,
-  },
-})
-
--- LSPs not in mason
-vim.lsp.enable({ "jarl" })
-
-return {
-  -- lspconfig
-  {
-    "https://github.com/neovim/nvim-lspconfig",
-    event = { "BufReadPre", "BufNewFile" },
-    dependencies = {
-      "https://github.com/mason-org/mason.nvim",
-      "https://github.com/mason-org/mason-lspconfig.nvim",
-    },
-    keys = {
-      { "gl", vim.diagnostic.open_float, desc = "Line Diagnostics" },
-      { "grx", vim.lsp.codelens.run, desc = "Run Codelens" },
-      { "<leader>cl", "<cmd>checkhealth vim.lsp<cr>", desc = "Lsp Info" },
-      {
-        "<leader>cD",
-        function()
-          Snacks.picker.diagnostics()
-        end,
-        desc = "Diagnostics (Workspace)",
-      },
-      {
-        "<leader>cd",
-        function()
-          Snacks.picker.diagnostics_buffer()
-        end,
-        desc = "Diagnostics (Document)",
-      },
-      {
-        "gd",
-        function()
-          Snacks.picker.lsp_definitions()
-        end,
-        desc = "Goto Definition",
-      },
-      {
-        "grr",
-        function()
-          Snacks.picker.lsp_references()
-        end,
-        desc = "Goto References",
-      },
-      {
-        "gD",
-        function()
-          Snacks.picker.lsp_declarations()
-        end,
-        desc = "Goto Declaration",
-      },
-      {
-        "gri",
-        function()
-          Snacks.picker.lsp_implementations()
-        end,
-        desc = "Goto Implementation",
-      },
-      {
-        "grt",
-        function()
-          Snacks.picker.lsp_type_definitions()
-        end,
-        desc = "Goto Type Definition",
-      },
-      {
-        "gO",
-        function()
-          Snacks.picker.lsp_symbols()
-        end,
-        desc = "Goto Document Symbols",
-      },
-      {
-        "<leader>cs",
-        function()
-          Snacks.picker.lsp_symbols()
-        end,
-        desc = "LSP Symbols",
-      },
-      {
-        "<leader>cS",
-        function()
-          Snacks.picker.lsp_workspace_symbols()
-        end,
-        desc = "LSP Workspace Symbols",
-      },
-      { "gK", vim.lsp.buf.signature_help, desc = "Signature Help" },
-      { "gra", vim.lsp.buf.code_action, desc = "Code Action", mode = { "n", "v" } },
-      { "<leader>ca", vim.lsp.buf.code_action, desc = "Code Action", mode = { "n", "v" } },
-    },
-    opts = {
-      servers = {
-        ansiblels = {},
-        bashls = {},
-        dockerls = {},
-        docker_compose_language_service = {},
-        marksman = {},
-        r_language_server = {},
-        lua_ls = {
-          settings = {
-            Lua = {
-              runtime = {
-                version = "LuaJIT",
-              },
-              workspace = {
-                library = {
-                  vim.env.VIMRUNTIME,
-                },
-                checkThirdParty = false,
-              },
-              completion = {
-                callSnippet = "Replace",
-              },
-              format = {
-                enable = false,
-              },
-            },
-          },
-        },
-      },
-    },
-    config = function(_, opts)
-      -- Prefer LSP folding if client supports it
-      vim.api.nvim_create_autocmd("LspAttach", {
-        callback = function(args)
-          local client = vim.lsp.get_client_by_id(args.data.client_id)
-          ---@diagnostic disable-next-line: need-check-nil
-          if client:supports_method("textDocument/foldingRange") then
-            local win = vim.api.nvim_get_current_win()
-            vim.wo[win][0].foldexpr = "v:lua.vim.lsp.foldexpr()"
-          end
-        end,
-      })
-
-      -- show colors as virtual text
-      vim.lsp.document_color.enable(true, nil, { style = "virtual" })
-
-      local servers = opts.servers
-      for server, settings in pairs(servers) do
-        vim.lsp.config(server, settings)
-        vim.lsp.enable(server)
-      end
-      -- get all the servers that are available through mason-lspconfig
-      local have_mason, mlsp = pcall(require, "mason-lspconfig")
-
-      local ensure_installed = vim.tbl_keys(opts.servers or {})
-      for server, _ in pairs(servers) do
-        ensure_installed[#ensure_installed + 1] = server
-      end
-
-      if have_mason then
-        mlsp.setup({ ensure_installed = ensure_installed })
-      end
-    end,
-  },
-
-  -- cmdline tools and lsp servers
-  {
-
-    "https://github.com/mason-org/mason.nvim",
-    cmd = "Mason",
-    keys = { { "<leader>cm", "<cmd>Mason<cr>", desc = "Mason" } },
-    build = ":MasonUpdate",
-    opts = {
-      ensure_installed = {
-        "air",
-        "shfmt",
-        "stylua",
-        "ansible-lint",
-        "mdformat",
-        "markdown-toc",
-        "tree-sitter-cli",
-      },
-      ui = {
-        icons = {
-          package_installed = "✓",
-          package_pending = "➜",
-          package_uninstalled = "✗",
-        },
-      },
-    },
-    config = function(_, opts)
-      require("mason").setup(opts)
-      local mr = require("mason-registry")
-      local function ensure_installed()
-        for _, tool in ipairs(opts.ensure_installed) do
-          local p = mr.get_package(tool)
-          if not p:is_installed() then
-            p:install()
-          end
-        end
-      end
-      if mr.refresh then
-        mr.refresh(ensure_installed)
-      else
-        ensure_installed()
-      end
-    end,
-  },
-}
