--- nvim/.config/nvim/lua/plugins/util.lua
@@ -1,340 +1,0 @@
-return {
-  -- collection of util plugins
-  {
-    "https://github.com/folke/snacks.nvim",
-    priority = 1000,
-    lazy = false,
-    keys = {
-      -- Pickers
-      -- core
-      {
-        "<leader>/",
-        function()
-          Snacks.picker.smart()
-        end,
-        desc = "Smart Find Files",
-      },
-      {
-        "<leader>:",
-        function()
-          Snacks.picker.command_history()
-        end,
-        desc = "Command History",
-      },
-      {
-        "<leader>n",
-        function()
-          Snacks.picker.notifications()
-        end,
-        desc = "Notification History",
-      },
-      -- Files
-      {
-        "<leader>fc",
-        function()
-          Snacks.picker.files({ cwd = vim.fn.stdpath("config"), title = "Config Files" })
-        end,
-        desc = "Find Config File",
-      },
-      {
-        "<leader>ff",
-        function()
-          Snacks.picker.files({
-            hidden = true,
-            cwd = vim.lsp.buf.list_workspace_folders()[1] or vim.fn.expand("%:h"),
-          })
-        end,
-        desc = "Find Files",
-      },
-      {
-        "<leader>fF",
-        function()
-          Snacks.picker.files({ cwd = vim.fn.expand("%:h") })
-        end,
-        desc = "Find Files (CWD)",
-      },
-      {
-        "<leader>fg",
-        function()
-          Snacks.picker.grep({ hidden = true, layout = { preset = "ivy" } })
-        end,
-        desc = "Grep",
-      },
-      {
-        "<leader>fr",
-        function()
-          Snacks.picker.recent()
-        end,
-        desc = "Recent",
-      },
-      {
-        "<leader>fu",
-        function()
-          Snacks.picker.undo()
-        end,
-        desc = "Undo History",
-      },
-      -- Buffers
-      {
-        "<leader>bb",
-        function()
-          Snacks.picker.buffers()
-        end,
-        desc = "Buffers",
-      },
-      -- search
-      {
-        "<leader>sk",
-        function()
-          Snacks.picker.keymaps()
-        end,
-        desc = "Keymaps",
-      },
-      {
-        "<leader>sc",
-        function()
-          Snacks.picker.colorschemes()
-        end,
-        desc = "Colorscheme",
-      },
-      {
-        "<leader>sC",
-        function()
-          Snacks.picker.commands()
-        end,
-        desc = "Commands",
-      },
-      {
-        "<leader>sh",
-        function()
-          Snacks.picker.help()
-        end,
-        desc = "Help Pages",
-      },
-      {
-        "<leader>sp",
-        function()
-          Snacks.picker.projects()
-        end,
-        desc = "Projects",
-      },
-      {
-        "<leader>se",
-        function()
-          Snacks.picker.icons()
-        end,
-        desc = "Emoji/Icons",
-      },
-      {
-        "<leader>sm",
-        function()
-          Snacks.picker.man()
-        end,
-        desc = "Man Pages",
-      },
-      {
-        "<leader><return>",
-        function()
-          Snacks.picker.resume()
-        end,
-        desc = "Resume Picker",
-      },
-      {
-        "<leader>sw",
-        function()
-          Snacks.picker.grep_word()
-        end,
-        desc = "Visual selection or word",
-        mode = { "n", "x" },
-      },
-      -- Other
-      {
-        "<leader>gg",
-        function()
-          Snacks.lazygit({ cwd = vim.lsp.buf.list_workspace_folders()[1] })
-        end,
-        desc = "Lazygit",
-      },
-      {
-        "<leader>gB",
-        function()
-          Snacks.gitbrowse()
-        end,
-        desc = "Gitbrowse",
-      },
-      {
-        "<leader>.",
-        function()
-          Snacks.scratch()
-        end,
-        desc = "Toggle Scratch Buffer",
-      },
-      {
-        [[<c-\>]],
-        mode = { "n", "t" },
-        function()
-          Snacks.terminal.toggle(nil, { cwd = vim.lsp.buf.list_workspace_folders()[1] or vim.fn.expand("%:h") })
-        end,
-        desc = "Toggle Terminal",
-      },
-      {
-        "]]",
-        function()
-          Snacks.words.jump(vim.v.count1)
-        end,
-        desc = "Next Reference",
-        mode = { "n", "t" },
-      },
-      {
-        "[[",
-        function()
-          Snacks.words.jump(-vim.v.count1)
-        end,
-        desc = "Prev Reference",
-        mode = { "n", "t" },
-      },
-    },
-    opts = {
-      bigfile = { enabled = false },
-      dashboard = {
-        preset = {
-          header = [[
-       ___       __   ___  ___  ________  _________       ___  ___  ________  ___  ___  ___
-      |\  \     |\  \|\  \|\  \|\   __  \|\___   ___\    |\  \|\  \|\   __  \|\  \|\  \|\  \
-      \ \  \    \ \  \ \  \\\  \ \  \|\  \|___ \  \_|    \ \  \\\  \ \  \|\  \ \  \ \  \ \  \
-       \ \  \  __\ \  \ \   __  \ \   __  \   \ \  \      \ \  \\\  \ \   ____\ \  \ \  \ \  \
-        \ \  \|\__\_\  \ \  \ \  \ \  \ \  \   \ \  \      \ \  \\\  \ \  \___|\ \__\ \__\ \__\
-         \ \____________\ \__\ \__\ \__\ \__\   \ \__\      \ \_______\ \__\    \|__|\|__|\|__|
-          \|____________|\|__|\|__|\|__|\|__|    \|__|       \|_______|\|__|        ___  ___  ___
-                                                                                   |\__\|\__\|\__\
-                                                                                   \|__|\|__|\|__|
-   ]],
-          keys = {
-            {
-              icon = " ",
-              key = "f",
-              desc = "Find File",
-              action = ":lua Snacks.dashboard.pick('files')",
-            },
-            { icon = " ", key = "n", desc = "New File", action = ":ene | startinsert" },
-            { icon = " ", key = "g", desc = "Find Text", action = ":lua Snacks.dashboard.pick('live_grep')" },
-            { icon = " ", key = "r", desc = "Recent Files", action = ":lua Snacks.dashboard.pick('oldfiles')" },
-            {
-              icon = " ",
-              key = "c",
-              desc = "Config",
-              action = ":lua Snacks.dashboard.pick('files', {cwd = vim.fn.stdpath('config')})",
-            },
-            { icon = " ", key = "p", desc = "Projects", action = ":lua Snacks.dashboard.pick('projects')" },
-            { icon = "󰒲 ", key = "L", desc = "Lazy", action = ":Lazy", enabled = package.loaded.lazy ~= nil },
-            { icon = " ", key = "q", desc = "Quit", action = ":qa" },
-          },
-        },
-        sections = {
-          { section = "header" },
-          { section = "keys", gap = 1, padding = 1 },
-          { section = "startup", icon = "💪 " },
-        },
-      },
-      indent = { enabled = true },
-      input = { enabled = true },
-      picker = {
-        win = {
-          input = {
-            keys = {
-              ["<a-s>"] = { "flash_jump", mode = { "n", "i" } },
-              ["s"] = { "flash_jump" },
-            },
-          },
-        },
-        actions = {
-          flash_jump = function(picker)
-            require("flash").jump({
-              pattern = "^",
-              label = { after = { 0, 0 } },
-              search = {
-                mode = "search",
-                exclude = {
-                  function(win)
-                    return vim.bo[vim.api.nvim_win_get_buf(win)].filetype ~= "snacks_picker_list"
-                  end,
-                },
-              },
-              action = function(match)
-                local idx = picker.list:row2idx(match.pos[1])
-                picker.list:_move(idx, true, true)
-              end,
-            })
-          end,
-        },
-      },
-      notifier = { timeout = 2000 },
-      quickfile = { enabled = true },
-      scope = { enabled = false },
-      scroll = { enabled = false },
-      statuscolumn = {
-        left = { "git" },
-        right = { "sign", "mark", "fold" },
-        folds = {
-          open = true,
-          git_hl = false,
-        },
-        git = {
-          patterns = { "GitSign", "MiniDiffSign" },
-        },
-        refresh = 50,
-      },
-      words = {},
-      styles = {
-        notification = {
-          wo = {
-            winblend = 10,
-          },
-        },
-        -- used by pickers
-        float = {
-          wo = {
-            winblend = 10,
-          },
-        },
-      },
-    },
-    init = function()
-      vim.api.nvim_create_autocmd("User", {
-        pattern = "VeryLazy",
-        callback = function()
-          -- Create some toggle mappings
-          Snacks.toggle.option("spell", { name = "Spelling" }):map("<leader>us")
-          Snacks.toggle.option("wrap", { name = "Wrap" }):map("<leader>uw")
-          Snacks.toggle.option("relativenumber", { name = "Relative Number" }):map("<leader>uL")
-          Snacks.toggle.diagnostics():map("<leader>ud")
-          Snacks.toggle.line_number():map("<leader>ul")
-          Snacks.toggle
-            .option("conceallevel", { off = 0, on = vim.o.conceallevel > 0 and vim.o.conceallevel or 2 })
-            :map("<leader>uc")
-          Snacks.toggle.inlay_hints():map("<leader>uh")
-          Snacks.toggle.indent():map("<leader>ug")
-          -- Conform
-          Snacks.toggle
-            .new({
-              name = "Autoformat",
-              get = function()
-                return not vim.g.disable_autoformat
-              end,
-              set = function(state)
-                vim.g.disable_autoformat = not state
-              end,
-            })
-            :map("<leader>uf")
-        end,
-      })
-    end,
-  },
-
-  -- Highlight undo
-  {
-    "https://github.com/tzachar/highlight-undo.nvim",
-    keys = { "u", "<C-r>" },
-    opts = {},
-  },
-}
