+++ nvim/.config/nvim/plugin/nvim-tree.lua
@@ -1,0 +1,55 @@
+vim.pack.add({
+  "https://github.com/nvim-tree/nvim-tree.lua",
+})
+
+local function on_attach(bufnr)
+  local api = require("nvim-tree.api")
+  local function opts(desc)
+    return { desc = "nvim-tree: " .. desc, buffer = bufnr, noremap = true, silent = true, nowait = true }
+  end
+  api.config.mappings.default_on_attach(bufnr)
+
+  vim.keymap.set("n", "l", api.node.open.edit, opts("Open"))
+  vim.keymap.set("n", "h", api.node.navigate.parent_close, opts("Close Directory"))
+  vim.keymap.set("n", "H", api.tree.collapse_all, opts("Collapse"))
+end
+
+require("nvim-tree").setup({
+  on_attach = on_attach,
+  disable_netrw = true,
+  hijack_netrw = false,
+  hijack_cursor = true,
+  reload_on_bufenter = true,
+  renderer = {
+    highlight_git = true,
+    indent_markers = {
+      enable = true,
+    },
+  },
+  diagnostics = {
+    enable = true,
+    show_on_dirs = true,
+    icons = {
+      hint = "󰌶 ",
+      info = " ",
+      warning = " ",
+      error = " ",
+    },
+  },
+  sync_root_with_cwd = true,
+  respect_buf_cwd = true,
+  update_focused_file = {
+    enable = true,
+    update_root = true,
+  },
+  filters = {
+    custom = { "^.git$" },
+  },
+  view = {
+    width = 40,
+  },
+  trash = {
+    cmd = "trash",
+  },
+})
+vim.keymap.set("n", "<leader>e", "<cmd>NvimTreeToggle<CR>", { desc = "Explorer" })
